third_party/blink/web_tests/platform/generic/fast/css/attribute-selector-recursive-update-on-setAttribute-expected.txt

This test checks that attribute selectors are recursively re-evaluated when the attribute changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(toggle).color is 'rgb(0, 128, 0)'
PASS getComputedStyle(toggle).color is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
